# client/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : ""
  },
  "basePath" : "/",
  "schemes" : [ "http" ],
  "paths" : {
    "/individuals" : {
      "get" : {
        "tags" : [ "individuals" ],
        "summary" : "Find individual matching specified attribute.",
        "description" : "",
        "operationId" : "findIndividuals",
        "produces" : [ "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "entity",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "attribute",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "value",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/IndividualComposite"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "individuals" ],
        "summary" : "Create individual",
        "description" : "",
        "operationId" : "createIndividual",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/IndividualComposite"
          }
        }, {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/IndividualComposite"
            }
          }
        }
      }
    },
    "/individuals/CAS/{casId}" : {
      "get" : {
        "tags" : [ "individuals" ],
        "summary" : "Read individual by CAS ID",
        "description" : "",
        "operationId" : "readIndividualByCasId",
        "parameters" : [ {
          "name" : "casId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/IndividualComposite"
            }
          }
        }
      }
    },
    "/individuals/{nedId}" : {
      "get" : {
        "tags" : [ "individuals" ],
        "summary" : "Read individual by Ned ID",
        "description" : "",
        "operationId" : "readIndividual",
        "produces" : [ "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/IndividualComposite"
            }
          }
        }
      }
    },
...
    "/individuals/{nedId}/auth" : {
      "get" : {
        "tags" : [ "individuals" ],
        "summary" : "List auth record(s)",
        "description" : "",
        "operationId" : "getAuthRecord",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ]
      }
    },
    "/individuals/{nedId}/auth/checkpassword" : {
      "post" : {
        "tags" : [ "individuals" ],
        "summary" : "Validate password",
        "description" : "",
        "operationId" : "checkPassword",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Auth"
          }
        } ]
      }
    },
    "/individuals/{nedId}/auth/{authId}" : {
      "put" : {
        "tags" : [ "individuals" ],
        "summary" : "Update auth record",
        "description" : "",
        "operationId" : "updateAuthRecord",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "authId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Auth"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/Auth"
            }
          }
        }
      }
    },
    "/individuals/{nedId}/degrees" : {
      "get" : {
        "tags" : [ "individuals" ],
        "summary" : "List degrees",
        "description" : "",
        "operationId" : "getDegrees",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Degree"
              }
            }
          }
        }
      }
    },
...
    "/individuals/{nedId}/groups" : {
      "get" : {
        "tags" : [ "individuals" ],
        "summary" : "List groups",
        "description" : "",
        "operationId" : "getGroups",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Group"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "individuals" ],
        "summary" : "Create group",
        "description" : "",
        "operationId" : "createGroup",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Group"
          }
        }, {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/Group"
            }
          }
        }
      }
    },
    "/individuals/{nedId}/groups/{groupId}" : {
      "get" : {
        "tags" : [ "individuals" ],
        "summary" : "Read group",
        "description" : "",
        "operationId" : "getGroup",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/Group"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "individuals" ],
        "summary" : "Update group",
        "description" : "",
        "operationId" : "updateGroup",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Group"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/Group"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "individuals" ],
        "summary" : "Delete group",
        "description" : "",
        "operationId" : "deleteGroup",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ]
      }
    },
    "/individuals/{nedId}/individualprofiles" : {
      "get" : {
        "tags" : [ "individuals" ],
        "summary" : "List profiles",
        "description" : "",
        "operationId" : "getProfiles",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Individualprofile"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "individuals" ],
        "summary" : "Add profile",
        "description" : "",
        "operationId" : "addProfile",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Individualprofile"
          }
        }, {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/Individualprofile"
            }
          }
        }
      }
    },
    "/individuals/{nedId}/individualprofiles/{profileId}" : {
      "get" : {
        "tags" : [ "individuals" ],
        "summary" : "Read profile",
        "description" : "",
        "operationId" : "getProfile",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "profileId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/Individualprofile"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "individuals" ],
        "summary" : "Update a profile",
        "description" : "",
        "operationId" : "updateProfile",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "profileId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Individualprofile"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/Individualprofile"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "individuals" ],
        "summary" : "Delete a profile",
        "description" : "",
        "operationId" : "deleteProfile",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "profileId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ]
      }
    },
...
    "/individuals/{nedId}/relationships" : {
      "get" : {
        "tags" : [ "individuals" ],
        "summary" : "List relationships",
        "description" : "",
        "operationId" : "getRelationship",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/void"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "individuals" ],
        "summary" : "Create relationship",
        "description" : "",
        "operationId" : "createRelationship",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Relationship"
          }
        }, {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/Relationship"
            }
          }
        }
      }
    },
...
    "/individuals/{uidType}/{uidValue}" : {
      "get" : {
        "tags" : [ "individuals" ],
        "summary" : "Read individual by UID",
        "description" : "",
        "operationId" : "readIndividualByUid",
        "parameters" : [ {
          "name" : "uidType",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "uidValue",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/IndividualComposite"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "Auth" : {
      "properties" : {
        "email" : {
          "type" : "string"
        },
        "emailid" : {
          "type" : "integer",
          "format" : "int32"
        },
        "authid" : {
          "type" : "string"
        },
        "password" : {
          "type" : "string"
        },
        "passwordreset" : {
          "type" : "string",
          "format" : "byte"
        },
        "verificationtoken" : {
          "type" : "string"
        },
        "verified" : {
          "type" : "string",
          "format" : "byte"
        },
        "isactive" : {
          "type" : "string",
          "format" : "byte"
        },
        "source" : {
          "type" : "string"
        },
        "nedid" : {
          "type" : "integer",
          "format" : "int32"
        },
        "sourcetypeid" : {
          "type" : "integer",
          "format" : "int32"
        },
        "id" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
...
    "IndividualComposite" : {
      "properties" : {
        "individualprofiles" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Individualprofile"
          }
        },
        "groups" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Group"
          }
        },
        "addresses" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Address"
          }
        },
        "emails" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Email"
          }
        },
        "phonenumbers" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Phonenumber"
          }
        },
        "uniqueidentifiers" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Uniqueidentifier"
          }
        },
        "degrees" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Degree"
          }
        },
        "urls" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Url"
          }
        },
        "relationships" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Relationship"
          }
        },
        "credentials" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Auth"
          }
        },
        "typeName" : {
          "type" : "string"
        },
        "asMap" : {
          "$ref" : "#/definitions/Map[Class,List[? extends org.plos.namedentity.api.entity.Entity]]"
        }
      }
    },
...
  }
}